requirements.txt
streamlit
pandas
numpy
yfinance
plotly
schedule
requests
python-dotenv
extra-streamlit-components
